[77] Update domain with home module
