+++ app/src/main/java/co/nimblehq/compose/crypto/ui/uimodel/CoinItemUiModel.kt
@@ -1,6 +1,6 @@
 package co.nimblehq.compose.crypto.ui.uimodel
 
-import co.nimblehq.compose.crypto.domain.model.CoinItem
+import co.nimblehq.compose.crypto.domain.home.model.CoinItem
 import java.math.BigDecimal
 
 data class CoinItemUiModel(
