+++ app/src/test/java/co/nimblehq/compose/crypto/test/MockUtil.kt
@@ -1,7 +1,7 @@
 package co.nimblehq.compose.crypto.test
 
 import co.nimblehq.compose.crypto.domain.model.CoinDetail
-import co.nimblehq.compose.crypto.domain.model.CoinItem
+import co.nimblehq.compose.crypto.domain.home.model.CoinItem
 import java.math.BigDecimal
 
 object MockUtil {
