+++ data/src/main/java/co/nimblehq/compose/crypto/data/home/model/response/CoinItemResponse.kt
@@ -1,7 +1,7 @@
 package co.nimblehq.compose.crypto.data.home.model.response
 
 import co.nimblehq.compose.crypto.data.extension.orZero
-import co.nimblehq.compose.crypto.domain.model.CoinItem
+import co.nimblehq.compose.crypto.domain.home.model.CoinItem
 import com.squareup.moshi.Json
 import java.math.BigDecimal
