+++ domain/src/main/java/co/nimblehq/compose/crypto/domain/home/HomeRepository.kt
@@ -1,5 +1,6 @@
 package co.nimblehq.compose.crypto.domain.home
 
+import co.nimblehq.compose.crypto.domain.home.model.CoinItem
 import co.nimblehq.compose.crypto.domain.model.*
 import kotlinx.coroutines.flow.Flow
