+++ domain/src/main/java/co/nimblehq/compose/crypto/domain/home/model/CoinItem.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.compose.crypto.domain.model
+package co.nimblehq.compose.crypto.domain.home.model
 
 import java.math.BigDecimal
